Move OkHttpClientFactory and TLSSocketFactory to sync package.
