+++ main/src/main/java/com/google/android/apps/muzei/sync/TLSSocketFactory.java
@@ -1,8 +1,20 @@
-package com.google.android.apps.muzei;
-
-/**
- * Created by marcopar on 11/02/17.
+/*
+ * Copyright 2014 Google Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
+
+package com.google.android.apps.muzei.sync;
 
 import java.io.IOException;
 import java.net.InetAddress;
